requirements.txt
tensorflow
mediapipe
h5py
scipy
scikit-learn
matplotlib
numpy
pandas
protobuf
google-generativeai
python-dotenv
flask
openai
gradio
moviepy
numpy
opencv-python
torch
